refactor: initial project architecture and structure update against circular imports
