+++ wrapper/configs/argument_config.py
@@ -3,10 +3,9 @@ import platform
 from typing import Optional
 from pydantic import BaseModel
 
-import wrapper.tools.messages as msg
-import wrapper.tools.commands as ccmd
-
-from wrapper.configs.directory_config import DirectoryConfig as dcfg
+from wrapper.tools import commands as ccmd
+from wrapper.utils import messages as msg
+from wrapper.configs import DirectoryConfig as dcfg
 
 
 class ArgumentConfig(BaseModel):
