+++ wrapper/modules/__init__.py
@@ -1,0 +1,2 @@
+from .kernel_builder import KernelBuilder
+from .assets_collector import AssetsCollector
