+++ wrapper/tools/__init__.py
@@ -1,4 +1,3 @@
 from .cleaning import remove, git, root
 from .commands import launch
 from .fileoperations import ucopy, download
-from .messages import banner, note, error, cancel, done, outputstream
