+++ wrapper/utils/resource_manager.py
@@ -3,12 +3,9 @@ import json
 import tarfile
 from typing import Optional
 
-import wrapper.tools.cleaning as cm
-import wrapper.tools.messages as msg
-import wrapper.tools.commands as ccmd
-import wrapper.tools.fileoperations as fo
-
-from wrapper.configs.directory_config import DirectoryConfig as dcfg
+from wrapper.tools import cleaning as cm, commands as ccmd, fileoperations as fo
+from wrapper.utils import messages as msg
+from wrapper.configs import DirectoryConfig as dcfg
 
 
 class ResourceManager:
